+++ test/blackbox-tests/test-cases/variants/module-fields/test.ml
@@ -6,14 +6,19 @@ let write_file ~file ~contents =
   output_string ch contents;
   close_out_noerr ch
 
-let lib_stanza ~modules_without_implementation ~virtual_modules =
-  sprintf "(library\n (name foo)%s%s)"
+let lib_stanza ~modules_without_implementation ~virtual_modules
+      ~private_modules =
+  sprintf "(library\n (name foo)%s%s%s)"
     (if modules_without_implementation then
        "\n (modules_without_implementation m)"
      else
        "")
     (if virtual_modules then
        "\n (virtual_modules m)"
+     else
+       "")
+    (if private_modules then
+       "\n (private_modules m)"
      else
        "")
 
@@ -28,11 +33,14 @@ let chdir dir ~f =
   end;
   Sys.chdir old_dir
 
-let gen_test ~impl ~modules_without_implementation ~virtual_modules =
-  printf "impl: %b. modules_without_implementation: %b. virtual_modules: %b\n%!"
-    impl modules_without_implementation virtual_modules;
+let gen_test ~impl ~modules_without_implementation ~virtual_modules
+  ~private_modules =
+  printf "impl: %b. modules_without_implementation: %b. \
+          virtual_modules: %b. private_modules: %b\n%!"
+    impl modules_without_implementation virtual_modules private_modules;
   let dir =
-    sprintf "%b-%b-%b" impl modules_without_implementation virtual_modules
+    sprintf "%b-%b-%b-%b" impl modules_without_implementation virtual_modules
+      private_modules
   in
   let _ = Sys.command (sprintf "mkdir -p %s" dir) in
   chdir dir ~f:(fun () ->
@@ -43,7 +51,8 @@ let gen_test ~impl ~modules_without_implementation ~virtual_modules =
     if impl then
       write_file ~file:"m.ml" ~contents:"";
     write_file ~file:"dune" ~contents:(
-      lib_stanza ~modules_without_implementation ~virtual_modules);
+      lib_stanza ~modules_without_implementation ~virtual_modules
+        ~private_modules);
     ignore (Sys.command "dune build");
     print_endline "-------------------------"
   )
@@ -51,10 +60,13 @@ let gen_test ~impl ~modules_without_implementation ~virtual_modules =
 let bools = [true; false]
 
 let () =
-  List.iter bools ~f:(fun impl ->
-    List.iter bools ~f:(fun modules_without_implementation ->
-      List.iter bools ~f:(fun virtual_modules ->
-        gen_test ~impl ~modules_without_implementation ~virtual_modules
+  List.iter bools ~f:(fun private_modules ->
+    List.iter bools ~f:(fun impl ->
+      List.iter bools ~f:(fun modules_without_implementation ->
+        List.iter bools ~f:(fun virtual_modules ->
+          gen_test ~impl ~modules_without_implementation ~virtual_modules
+            ~private_modules
+        )
       )
     )
   )